restructure(Thread): Move ThreadPost to containers
