+++ src/modules/Thread/components/Comment.jsx
