+++ src/modules/Thread/components/index.js
@@ -1,1 +1,8 @@
 export {default as Overlay} from './Overlay'
+export {default as Comment} from './Comment'
+export {default as ControlMenu} from './ControlMenu'
+export {default as MediaInfo} from './MediaInfo'
+export {default as References} from './References'
+export {default as Title} from './Title'
+
+export * from './Controllers'
