+++ src/modules/Thread/containers/index.js
@@ -1,0 +1,3 @@
+export {default as ThreadPost} from './ThreadPost';
+export {default as ThreadMedia} from './ThreadMedia';
+export {default as ThreadControls} from './ThreadControls';
